+++ libft/ft_putnbr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wrhett <wrhett@student.42.fr>              +#+  +:+       +#+        */
+/*   By: wrhett <wrhett@student.21-school.ru>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/09 12:19:46 by wrhett            #+#    #+#             */
-/*   Updated: 2019/09/18 15:39:40 by wrhett           ###   ########.fr       */
+/*   Updated: 2020/05/31 18:29:21 by wrhett           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	ft_putnbr(int nb)
 {
 	unsigned int num;
 
-	num = (nb < 0) ? (unsigned int)-nb : nb;
+	num = (nb < 0) ? (unsigned int)-nb : (unsigned int)nb;
 	if (nb < 0)
 		ft_putchar('-');
 	if (num >= 10)
